#1371 活動記録の編集・削除・マイページでの一覧表示：レビュー指摘事項修正

+++ app/Http/Controllers/ActivitiesController.php
@@ -59,9 +59,11 @@ class ActivitiesController extends Controller
         $activity = Activity::findOrFail($id);
         $user = \Auth::user();
 
-        \Storage::disk('public')->delete($activity->image);
-        $path = $request->file('image')->store('shop-images', 'public');
-        $activity->image = $path;
+        if ($request->hasFile('image')) {
+            \Storage::disk('public')->delete($activity->image);
+            $path = $request->file('image')->store('shop-images', 'public');
+            $activity->image = $path;
+        }
 
         $activity->user_id = $user->id;
         $activity->shop_name = $request->shop_name;
